+++ cedar-policy/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Move public API for partial evaluation behind experimental feature flag.
+- Added an option to eagerly evaluate entity attributes and re-use across calls to `is_authorized`
 
 ### Added
 
@@ -11,6 +12,7 @@
   - `namespace_components` to get the namespace as an iterator over its components.
   - `namespace` to get the namespace as a single string.
 - Fixed bug (#150) around implicit namespaces in action definitions.
+- Support `Request`s with `Unknown` fields for partial evaluation.
 
 ### Changed